Defect 14498: [Gridx_FVT]Focus: Focus jump to next row when rearrange row by CTRL+Arrow key.
Refine the 'if' condition for moving focus.
In general, pressing up/down arrow key will move focus to upper/lower
cell. But when dnd/row module is added and ctrl key is pressed at the
same time, it should move the selected rows up or down. The focus should
remain on the same cell.